Created RecordIndexer class for cleaner code [skip ci]
